ajout d'un canevas pour le nouveau prefabs "Cellule" (oublie)
Les vies "curent_pv" sont mainteant des float et les toxines font moins
de dégat.
Les cellules sont maintenant infecable et sensible aux toxines.
Le joueur est lui aussi sensible aux toxines

+++ InsideInvaders/Assets/Components/BarreDeVie.cs
@@ -3,6 +3,6 @@
 public class BarreDeVie : MonoBehaviour {
 	// Advice: FYFY component aims to contain only public members (according to Entity-Component-System paradigm).
 	public int max_pv = 200;
-	public int current_pv = 200;
+	public float current_pv = 200;
 
 }
